+++ pytorch_lightning/callbacks/quantization.py
@@ -75,16 +75,14 @@ def _recursive_hasattr(obj, attribs, state=True):
         attrib, attribs = attribs.split('.', 1)
         if hasattr(obj, attrib):
             return _recursive_hasattr(getattr(obj, attrib), attribs, state)
-        else:
-            return False
-    else:
-        return state & hasattr(obj, attribs)
+        return False
+    return state and hasattr(obj, attribs)
 
 
 class QuantizationAwareTraining(Callback):
     """
-    Quantization allows speed-up inference and decrease model size by performing computations and storing tensors
-     at lower bitwidths than floating point precision such as INT8 or FLOAT16.
+    Quantization allows speeding up inference and decreasing memory requirements by performing computations and storing tensors
+     at lower bitwidths (such as INT8 or FLOAT16) than floating point precision.
     We use native PyTorch API so for more information see
      `Quantization <https://pytorch.org/docs/stable/quantization.html#quantization-aware-training>_`
     """
@@ -104,25 +102,25 @@ class QuantizationAwareTraining(Callback):
             qconfig: define quantization configuration see: `torch.quantization.QConfig
              <https://pytorch.org/docs/stable/torch.quantization.html?highlight=qconfig#torch.quantization.QConfig>_`
                 or use pre-defined: 'fbgemm' for server inference and 'qnnpack' for mobile inference
-            observer_type: switching between `MovingAverageMinMaxObserver` as "average" (default)
-                and `HistogramObserver` as "histogram" which is more computational costly
-            collect_quantization: define custom count or function when you shall collect quantization statistic
+            observer_type: allows switching between `MovingAverageMinMaxObserver` as "average" (default)
+                and `HistogramObserver` as "histogram" which is more computationally expensive
+            collect_quantization: count or custom function to collect quantization statistics
 
-                - with default ``None`` you call the quantization observer for each module forward,
+                - with default ``None`` the quantization observer is called each module forward,
                     typical use-case can be reflecting user augmentation
-                - custom call count to limit explicit number of calls starting from beginning
-                - custom lambda function with single trainer argument,
+                - custom call count to set a fixed number of calls, starting from the beginning
+                - custom ``Callable`` function with single trainer argument,
                     see example when you limit call only for last epoch::
 
                     def custom_trigger_last(trainer):
                         return trainer.current_epoch == (trainer.max_epochs - 1)
 
-                    QuantizationAwareTraining(lambda_trigger=custom_trigger_last)
+                    QuantizationAwareTraining(collect_quantization=custom_trigger_last)
 
-            modules_to_fuse: allows you fuse a few layer together as you can see in `diagram
+            modules_to_fuse: allows you fuse a few layers together as shown in `diagram
              <https://pytorch.org/docs/stable/quantization.html#quantization-aware-training>_`
-                to find what layer types can be fue=sed check https://github.com/pytorch/pytorch/pull/43286
-            input_compatible: preserve quent/dequant layers which allows to feat any input as to the original model,
+                to find which layer types can be fused, check https://github.com/pytorch/pytorch/pull/43286
+            input_compatible: preserve quant/dequant layers. This allows to feat any input as to the original model,
                 but break compatibility to torchscript
         """
         if not isinstance(qconfig, (str, QConfig)):
@@ -137,7 +135,7 @@ class QuantizationAwareTraining(Callback):
 
         if collect_quantization is not None and not isinstance(collect_quantization, (int, Callable)):
             raise MisconfigurationException(
-                f'Unsupported `collect_quantization` "{observer_type}", allowed are int or callable.'
+                f'Unsupported `collect_quantization` "{collect_quantization}", allowed are `int` or `Callable`.'
             )
         self._collect_quantization = collect_quantization
 
@@ -177,8 +175,7 @@ class QuantizationAwareTraining(Callback):
         elif isinstance(self._qconfig, QConfig):
             pl_module.qconfig = self._qconfig
 
-        if self.modules_to_fuse:
-            self._check_feasible_fuse(pl_module)
+        if self._check_feasible_fuse(pl_module)
             torch.quantization.fuse_modules(pl_module, self.modules_to_fuse, inplace=True)
 
         # Prepare the model for QAT. This inserts observers and fake_quants in